manual_requirements.txt
python (conda create -n ecg_xai python=3.10 -y)
flake8 (conda install flake8 -y)
yapf (conda install yapf -y)

numba (conda install numba -y)

pytorch (conda install pytorch torchvision torchaudio pytorch-cuda=11.7 -c pytorch -c nvidia -y)
- On Mac: conda install pytorch torchvision torchaudio -c pytorch -y
pytorch lightning(conda install pytorch-lightning -c conda-forge -y)
torchmetrics (conda install -c conda-forge torchmetrics -y)
optuna (conda install -c conda-forge optuna -y)
tensorboard (conda install -c conda-forge tensorboard -y)

seaborn (conda install seaborn -y)

if seaborn not installed
- pandas (conda install pandas -y)
- matplotlib (conda install matplotlib -y)


neurokit2 (conda install neurokit2 -c conda-forge -y)
- if using 'cwt' (continuous wavelet transform) in neurokit2, install pywavelets (conda install pywavelets -y)
- intel accelerated scikit-learn (conda install scikit-learn-intelex -y)
wfdb (conda install wfdb -c conda-forge -y)
ecg_plot (pip install ecg_plot)